requirements.txt
alembic==1.4.1
astroid==2.3.3
bcrypt==3.1.7
cffi==1.14.0
click==6.7
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
isort==4.3.21
itsdangerous==0.24
Jinja2==2.10.1
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.0
mccabe==0.6.1
psycopg2==2.8.4
pycparser==2.20
PyJWT==1.4.2
pylint==2.4.4
python-dateutil==2.8.1
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.13
typed-ast==1.4.1
Werkzeug==0.15.5
wrapt==1.11.2
